// frontend/src/api/generated/models/stocktake-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { StocktakeStatus } from './stocktake-status';

/**
 * 
 * @export
 * @interface StocktakeSummary
 */
export interface StocktakeSummary {
    /**
     * 棚卸ヘッダID
     * @type {number}
     * @memberof StocktakeSummary
     */
    'id': number;
    /**
     * 棚卸番号
     * @type {string}
     * @memberof StocktakeSummary
     */
    'stocktakeNumber': string;
    /**
     * 倉庫ID
     * @type {number}
     * @memberof StocktakeSummary
     */
    'warehouseId': number;
    /**
     * 倉庫名
     * @type {string}
     * @memberof StocktakeSummary
     */
    'warehouseName': string;
    /**
     * 棚卸対象説明
     * @type {string}
     * @memberof StocktakeSummary
     */
    'targetDescription': string;
    /**
     * 
     * @type {StocktakeStatus}
     * @memberof StocktakeSummary
     */
    'status': StocktakeStatus;
    /**
     * 棚卸明細総件数
     * @type {number}
     * @memberof StocktakeSummary
     */
    'totalLines': number;
    /**
     * 実数入力済み件数
     * @type {number}
     * @memberof StocktakeSummary
     */
    'countedLines': number;
    /**
     * 棚卸開始日時
     * @type {string}
     * @memberof StocktakeSummary
     */
    'startedAt': string;
    /**
     * 棚卸開始者氏名
     * @type {string}
     * @memberof StocktakeSummary
     */
    'startedByName': string;
    /**
     * 棚卸確定日時
     * @type {string}
     * @memberof StocktakeSummary
     */
    'confirmedAt'?: string | null;
    /**
     * 棚卸確定者氏名
     * @type {string}
     * @memberof StocktakeSummary
     */
    'confirmedByName'?: string | null;
}