// frontend/src/api/generated/models/stocktake-summary-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { StocktakeSummary } from './stocktake-summary';

/**
 * 
 * @export
 * @interface StocktakeSummaryPageResponse
 */
export interface StocktakeSummaryPageResponse {
    /**
     * 
     * @type {Array<StocktakeSummary>}
     * @memberof StocktakeSummaryPageResponse
     */
    'content': Array<StocktakeSummary>;
    /**
     * 
     * @type {number}
     * @memberof StocktakeSummaryPageResponse
     */
    'page': number;
    /**
     * 
     * @type {number}
     * @memberof StocktakeSummaryPageResponse
     */
    'size': number;
    /**
     * 
     * @type {number}
     * @memberof StocktakeSummaryPageResponse
     */
    'totalElements': number;
    /**
     * 
     * @type {number}
     * @memberof StocktakeSummaryPageResponse
     */
    'totalPages': number;
}